% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calPlot}
\alias{calPlot}
\title{Calibration of the observed probability vs. prediction probability}
\usage{
calPlot(model, data, class)
}
\arguments{
\item{model}{The prediction model.}

\item{data}{A dataframe containing features of the test samples.}

\item{class}{Class of the test samples.}
}
\description{
https://rdrr.io/cran/caret/man/calibration.html
}
\examples{
calPlot()
}
