% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{cancerPrediction}
\alias{cancerPrediction}
\title{Predict the occurrence of tumor}
\usage{
cancerPrediction(file_path, model = NULL, rCNV = NULL,
  return = FALSE)
}
\description{
Predict the occurrence of tumor
}
\examples{
cancerPrediction()
}
