% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{cnv.cal}
\alias{cnv.cal}
\title{https://github.com/hliang/cnv-seq/blob/master/cnv/R/cnv.R}
\usage{
cnv.cal(data, log2.threshold = 0.6, chromosomal.normalization = FALSE,
  annotate = FALSE, minimum.window = 4)
}
\description{
https://github.com/hliang/cnv-seq/blob/master/cnv/R/cnv.R
}
\examples{
cnv.cal()
}
\keyword{internal}
