% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{dataMerge}
\alias{dataMerge}
\title{Merge two 'data' files}
\usage{
dataMerge(data_1, data_2)
}
\description{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4273479/pdf/13336_2014_Article_15.pdf
}
\examples{
dataMerge()
}
