% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc2data}
\alias{doc2data}
\title{Using 'doc' files to make the 'data' file}
\usage{
doc2data(doc.list)
}
\description{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4273479/pdf/13336_2014_Article_15.pdf
}
\examples{
doc2data()
}
