% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{featureSelection2}
\alias{featureSelection2}
\title{Cross platform feature selection method_2}
\usage{
featureSelection2(nonCancerListA, CancerListA, nonCancerListB, CancerListB,
  Cri = 0.2)
}
\description{
Cross platform feature selection method_2
}
\examples{
featureSelection2()
}
