% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCNViewer.R
\name{plotFrequency}
\alias{plotFrequency}
\title{Plot the frequency of CNV}
\usage{
plotFrequency(input, Cri = 0.05, return = TRUE, output_path = NULL,
  recCri = NULL)
}
\description{
Plot the frequency of CNV
}
\examples{
plotFrequency()
}
