% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSelectedFeatures}
\alias{plotSelectedFeatures}
\title{Quantsmooth plot of selected CNV features}
\usage{
plotSelectedFeatures(output_path, nonCancerListA, CancerListA, nonCancerListB,
  CancerListB, Cri = 0.2)
}
\description{
Quantsmooth plot of selected CNV features
}
\examples{
plotSelectedFeatures()
}
