% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc2data}
\alias{doc2data}
\title{Using 'doc' files to make the 'data' file}
\usage{
doc2data(doc.list, write.file.path = NULL)
}
\description{
Using 'doc' files to make the 'data' file
}
\examples{
doc2data()
}
