% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{featureSelection}
\alias{featureSelection}
\title{Cross platform feature selection}
\usage{
featureSelection(nonCancerListA, CancerListA, nonCancerListB, CancerListB,
  Cri = 0.2)
}
\description{
Cross platform feature selection
}
\examples{
featureSelection()
}
