% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictTumor}
\alias{predictTumor}
\title{Predict the tumor occurrence}
\usage{
predictTumor(file_path, model = NULL, rCNV = NULL, return = FALSE)
}
\description{
Predict the tumor occurrence
}
\examples{
predictTumor()
}
